Merge analytics from all language subdomains
This site runs on different subdomains for different languages (`de.communitycodeofconduct.com`, `it.communitycodeofconduct.com`, etc).

This PR uses a new capability in Plausible to override the domain for which traffic is measured. This means that all traffic will show up on the `communitycodeofconduct.com` dashboard, even if the visitor was actually on a subdomain.